pyproject.toml
[project]
name = "hiremeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.4",
    "fastapi>=0.135.2",
    "ipykernel>=7.1.0",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.5",
    "langsmith>=0.5.1",
    "nest-asyncio>=1.6.0",
    "openai>=2.14.0",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.26",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "streamlit>=1.55.0",
    "uvicorn>=0.44.0",
]
